requirements.txt
Flask
Flask-Sockets
Flask-RESTful
Flask-SQLAlchemy
PyJWT
uuid
gevent